Accept files name as command line argument and enable streaming
This commit changes the behavior of the program in such a way that it
accepts input and output filenames as command line arguments instead of
prompting for them. This makes usage in scripts easier.

Both arguments may be omitted or set to "-" in which case the json2nix
will read from stdin resp. write to stdout.