Fix wrap key in quotes if it contains special characters
There are many special characters that Nix treats specially and that may
cause trouble if found inside the key of an attribute set.

Eg. if a key contains a dot, Nix will interpret it as a nested attribute
set. Or if a key contains a slash, Nix recognizes it as Path. Also there
are semicolons, braces etc.

Thus, this commit introduces a condition to add quotes around the key if
any other sign apart from [a-zA-Z0-9-_] is found.